+++ src/test/kotlin/id/walt/vclib/vcs/CredentialsTests.kt
@@ -26,6 +26,8 @@ class CredentialsTests : StringSpec({
         File("src/test/resources/serialized/GaiaxSelfDescription.json").writeText(GaiaxSelfDescription.template?.invoke()!!.encode())
         File("src/test/resources/serialized/GaiaxServiceOffering.json").writeText(GaiaxServiceOffering.template?.invoke()!!.encode())
         File("src/test/resources/serialized/VerifiablePresentation.json").writeText(VerifiablePresentation.template?.invoke()!!.encode())
+        File("src/test/resources/serialized/VerifiableVaccinationCertificate.json").writeText(VerifiableVaccinationCertificate.template?.invoke()!!.encode())
+        File("src/test/resources/serialized/ProofOfResidence.json").writeText(ProofOfResidence.template?.invoke()!!.encode())
 
     }